website/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
django-filter==24.2
djangorestframework==3.15.2
gunicorn==22.0.0
joblib==1.4.2
Markdown==3.6
mysqlclient==2.2.4
numpy==2.0.0
packaging==24.1
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
scikit-learn==1.5.0
scipy==1.14.0
six==1.16.0
sqlparse==0.5.0
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
